+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/http/ZHttpServer.scala
@@ -2,6 +2,7 @@ package io.iohk.atala.agent.server.http
 
 import cats.implicits.*
 import io.iohk.atala.agent.server.http.ZHttpEndpoints
+import io.iohk.atala.api.http.BadRequest
 import io.iohk.atala.pollux.schema.SchemaRegistryServerEndpoints
 import io.iohk.atala.pollux.service.{SchemaRegistryService, SchemaRegistryServiceInMemory}
 import org.http4s.*
@@ -10,24 +11,30 @@ import org.http4s.server.Router
 import org.slf4j.LoggerFactory
 import sttp.model.StatusCode
 import sttp.tapir.server.ServerEndpoint
+import sttp.tapir.server.http4s.Http4sServerOptions
 import sttp.tapir.server.http4s.ztapir.ZHttp4sServerInterpreter
 import sttp.tapir.server.interceptor.log.DefaultServerLog
 import sttp.tapir.ztapir.ZServerEndpoint
-import zhttp.http.HttpApp
-import zhttp.service.server.ServerChannelFactory
-import zhttp.service.{EventLoopGroup, Server}
-import zio.interop.catz.*
 import zio.*
+import zio.interop.catz.*
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
 object ZHttp4sBlazeServer {
+
+  val serverOptions: Http4sServerOptions[Task] = Http4sServerOptions
+    .customiseInterceptors[Task]
+    .defaultHandlers(BadRequest.failureResponseHandler)
+    .options
+
   def start(
       endpoints: List[ZServerEndpoint[Any, Any]],
       port: Int
   ): Task[ExitCode] = {
     val http4sEndpoints: HttpRoutes[Task] =
-      ZHttp4sServerInterpreter().from(endpoints).toRoutes
+      ZHttp4sServerInterpreter(serverOptions)
+        .from(endpoints)
+        .toRoutes
 
     val serve: Task[Unit] =
       ZIO.executor.flatMap(executor =>