+++ tests/e2e-tests/src/test/kotlin/features/issue_credentials/IssueCredentialsSteps.kt
@@ -3,39 +3,14 @@ package features.issue_credentials
 import io.cucumber.java.en.Given
 import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
-import net.serenitybdd.rest.SerenityRest
-import net.serenitybdd.screenplay.Actor
-import net.serenitybdd.screenplay.rest.abilities.CallAnApi
 import net.serenitybdd.screenplay.rest.interactions.Get
 import net.serenitybdd.screenplay.rest.interactions.Post
 import net.serenitybdd.screenplay.rest.questions.ResponseConsequence
-import net.thucydides.core.util.EnvironmentVariables
 import api_models.Connection
 import api_models.Credential
-import io.restassured.path.json.JsonPath
-import org.awaitility.Awaitility
-import org.awaitility.core.ConditionTimeoutException
-import org.awaitility.kotlin.withPollInterval
-import org.awaitility.pollinterval.FixedPollInterval
-import java.time.Duration
+import extentions.WithAgents
 
-
-class IssueCredentialsSteps {
-
-    private lateinit var environmentVariables: EnvironmentVariables
-    private lateinit var acme: Actor
-    private lateinit var bob: Actor
-    private lateinit var lastResponse: JsonPath
-
-    @Given("Acme and Bob agents")
-    fun acmeAndBobAgents() {
-        val theRestApiBaseUrlIssuer = environmentVariables.optionalProperty("restapi.baseurl.issuer")
-            .orElse("http://localhost:8080")
-        val theRestApiBaseUrlHolder = environmentVariables.optionalProperty("restapi.baseurl.holder")
-            .orElse("http://localhost:8090")
-        acme = Actor.named("Acme").whoCan(CallAnApi.at(theRestApiBaseUrlIssuer))
-        bob = Actor.named("Bob").whoCan(CallAnApi.at(theRestApiBaseUrlHolder))
-    }
+class IssueCredentialsSteps : WithAgents() {
 
     @Given("Acme and Bob have an existing connection")
     fun acmeAndBobHaveAnExistingConnection() {
@@ -53,7 +28,7 @@ class IssueCredentialsSteps {
                 it.statusCode(201)
             }
         )
-        val acmeConnection = SerenityRest.lastResponse().jsonPath().getObject("", Connection::class.java)
+        val acmeConnection = lastResponse().getObject("", Connection::class.java)
 
         // Here out of band transfer of connection QR code is happening
         // and Bob (Holder) gets an invitation URL
@@ -73,56 +48,46 @@ class IssueCredentialsSteps {
                 it.statusCode(200)
             }
         )
-        val bobConnection = SerenityRest.lastResponse().jsonPath().getObject("", Connection::class.java)
+        val bobConnection = lastResponse().getObject("", Connection::class.java)
 
         // Acme(Issuer) checks their connections to check if invitation was accepted by Bob(Holder)
         // and sends final connection response
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    acme.attemptsTo(
-                        Get.resource("/connections/${acmeConnection.connectionId}")
-                    )
-                    acme.should(
-                        ResponseConsequence.seeThatResponse("Get issuer connections") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getObject("", Connection::class.java).state == "ConnectionResponseSent"
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Issuer did not sent final connection confirmation! Connection didn't reach ConnectionResponseSent state."
-            )
-        }
-        acme.remember("did", lastResponse.getObject("", Connection::class.java).myDid)
-        acme.remember("holderDid", lastResponse.getObject("", Connection::class.java).theirDid)
+        wait(
+            {
+                acme.attemptsTo(
+                    Get.resource("/connections/${acmeConnection.connectionId}"),
+                )
+                acme.should(
+                    ResponseConsequence.seeThatResponse("Get issuer connections") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse()
+                    .getObject("", Connection::class.java).state == "ConnectionResponseSent"
+            },
+            "Issuer did not sent final connection confirmation! Connection didn't reach ConnectionResponseSent state."
+        )
+        acme.remember("did", lastResponse().getObject("", Connection::class.java).myDid)
+        acme.remember("holderDid", lastResponse().getObject("", Connection::class.java).theirDid)
 
         // Bob (Holder) receives final connection response
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    bob.attemptsTo(
-                        Get.resource("/connections/${bobConnection.connectionId}")
-                    )
-                    bob.should(
-                        ResponseConsequence.seeThatResponse("Get holder connections") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getObject("", Connection::class.java).state == "ConnectionResponseReceived"
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Holder did not receive final connection confirmation! Connection didn't reach ConnectionResponseReceived state."
-            )
-        }
-        bob.remember("did", lastResponse.getObject("", Connection::class.java).myDid)
-        bob.remember("issuerDid", lastResponse.getObject("", Connection::class.java).theirDid)
+        wait(
+            {
+                bob.attemptsTo(
+                    Get.resource("/connections/${bobConnection.connectionId}")
+                )
+                bob.should(
+                    ResponseConsequence.seeThatResponse("Get holder connections") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse().getObject("", Connection::class.java).state == "ConnectionResponseReceived"
+            },
+            "Holder did not receive final connection confirmation! Connection didn't reach ConnectionResponseReceived state."
+        )
+
+        bob.remember("did", lastResponse().getObject("", Connection::class.java).myDid)
+        bob.remember("issuerDid", lastResponse().getObject("", Connection::class.java).theirDid)
         // Connection established. Both parties exchanged their DIDs with each other
     }
 
@@ -158,28 +123,23 @@ class IssueCredentialsSteps {
 
     @When("Bob requests the credential")
     fun bobRequestsTheCredential() {
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    bob.attemptsTo(
-                        Get.resource("/issue-credentials/records")
-                    )
-                    bob.should(
-                        ResponseConsequence.seeThatResponse("Credential records") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getList("items", Credential::class.java).findLast { it.protocolState == "OfferReceived" } != null
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Holder was unable to receive the credential offer from Issuer! Protocol state did not achieve OfferReceived state."
-            )
-        }
+        wait(
+            {
+                bob.attemptsTo(
+                    Get.resource("/issue-credentials/records")
+                )
+                bob.should(
+                    ResponseConsequence.seeThatResponse("Credential records") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse().getList("items", Credential::class.java).findLast { it.protocolState == "OfferReceived" } != null
+            },
+            "Holder was unable to receive the credential offer from Issuer! Protocol state did not achieve OfferReceived state."
+        )
 
-        val recordId = lastResponse.getList("items", Credential::class.java).findLast { it.protocolState == "OfferReceived" }!!.recordId
+        val recordId = lastResponse().getList("items", Credential::class.java)
+            .findLast { it.protocolState == "OfferReceived" }!!.recordId
         bob.remember("recordId", recordId)
 
         bob.attemptsTo(
@@ -194,27 +154,23 @@ class IssueCredentialsSteps {
 
     @When("Acme issues the credential")
     fun acmeIssuesTheCredential() {
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    acme.attemptsTo(
-                        Get.resource("/issue-credentials/records")
-                    )
-                    acme.should(
-                        ResponseConsequence.seeThatResponse("Credential records") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getList("items", Credential::class.java).findLast { it.protocolState == "RequestReceived" } != null
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Issuer was unable to receive the credential request from Holder! Protocol state did not achieve RequestReceived state."
-            )
-        }
-        val recordId = lastResponse.getList("items", Credential::class.java).findLast { it.protocolState == "RequestReceived" }!!.recordId
+        wait(
+            {
+                acme.attemptsTo(
+                    Get.resource("/issue-credentials/records")
+                )
+                acme.should(
+                    ResponseConsequence.seeThatResponse("Credential records") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse().getList("items", Credential::class.java)
+                    .findLast { it.protocolState == "RequestReceived" } != null
+            },
+            "Issuer was unable to receive the credential request from Holder! Protocol state did not achieve RequestReceived state."
+        )
+        val recordId = lastResponse().getList("items", Credential::class.java)
+            .findLast { it.protocolState == "RequestReceived" }!!.recordId
         acme.attemptsTo(
             Post.to("/issue-credentials/records/${recordId}/issue-credential")
         )
@@ -224,50 +180,39 @@ class IssueCredentialsSteps {
             }
         )
 
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    acme.attemptsTo(
-                        Get.resource("/issue-credentials/records/${recordId}")
-                    )
-                    acme.should(
-                        ResponseConsequence.seeThatResponse("Credential records") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getObject("", Credential::class.java).protocolState == "CredentialSent"
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Issuer was unable to issue the credential! Protocol state did not achieve CredentialSent state."
-            )
-        }
+        wait(
+            {
+                acme.attemptsTo(
+                    Get.resource("/issue-credentials/records/${recordId}")
+                )
+                acme.should(
+                    ResponseConsequence.seeThatResponse("Credential records") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse().getObject("", Credential::class.java).protocolState == "CredentialSent"
+            },
+            "Issuer was unable to issue the credential! Protocol state did not achieve CredentialSent state."
+        )
     }
 
     @Then("Bob has the credential issued")
     fun bobHasTheCredentialIssued() {
-        try {
-            Awaitility.await().withPollInterval(FixedPollInterval(Duration.ofSeconds(3L)))
-                .atMost(Duration.ofSeconds(60L))
-                .until {
-                    bob.attemptsTo(
-                        Get.resource("/issue-credentials/records/${bob.recall<String>("recordId")}")
-                    )
-                    bob.should(
-                        ResponseConsequence.seeThatResponse("Credential records") {
-                            it.statusCode(200)
-                        }
-                    )
-                    lastResponse = SerenityRest.lastResponse().jsonPath()
-                    lastResponse.getObject("", Credential::class.java).protocolState == "CredentialReceived"
-                }
-        } catch (err: ConditionTimeoutException) {
-            throw ConditionTimeoutException(
-                "Holder was unable to receive the credential from Issuer! Protocol state did not achieve CredentialReceived state."
-            )
-        }
-        println(lastResponse.getObject("items", Credential::class.java))
+        wait(
+            {
+                bob.attemptsTo(
+                    Get.resource("/issue-credentials/records/${bob.recall<String>("recordId")}")
+                )
+                bob.should(
+                    ResponseConsequence.seeThatResponse("Credential records") {
+                        it.statusCode(200)
+                    }
+                )
+                lastResponse().getObject("", Credential::class.java).protocolState == "CredentialReceived"
+            },
+            "Holder was unable to receive the credential from Issuer! Protocol state did not achieve CredentialReceived state."
+        )
+        val achievedCredential = lastResponse().getObject("", Credential::class.java)
+        println(achievedCredential)
     }
 }