+++ tests/e2e-tests/src/test/kotlin/runners/E2eTestsRunner.kt
@@ -5,16 +5,15 @@ import net.serenitybdd.cucumber.CucumberWithSerenity
 import org.junit.runner.RunWith
 
 @CucumberOptions(
-    plugin = [
-        "pretty",
-        "json:target/serenity-reports/cucumber_report.json",
-        "html:target/serenity-reports/cucumber_report.html"
-    ],
     features = [
         "src/test/resources/features"
     ],
     glue = ["features"],
-    snippets = CucumberOptions.SnippetType.CAMELCASE
+    snippets = CucumberOptions.SnippetType.CAMELCASE,
+    plugin = [
+        "pretty",
+        "json:target/serenity-reports/cucumber_report.json"
+    ],
 )
 @RunWith(CucumberWithSerenity::class)
 class E2eTestsRunner